#655: extract GHC.Base — compiler-magic types and core classes
Big content step of Milestone 1a.  Moves the type/class spine of the
Prelude into `lib/ghc-internal/GHC/Base.hs`, mirroring GHC's
`ghc-internal:GHC.Base`:

- `data Bool/Maybe/Either/Ordering` + `String` synonym
- Fixity declarations
- `error`, `putChar`/`putStr`/`putStrLn`
- Boolean (`not`/`&&`/`||`/`otherwise`)
- Monomorphic Int arithmetic (`(+)`, `(-)`, `(*)`, `negate`, `abs`,
  `div`, `mod`, `max`, `min`, `signum`, `even`, `odd`)
- `Eq`, `Ord`, `Bounded`, `Enum`, `Show` classes + primitive instances
- Show helpers (`showString`, `showLitChar`, `showListWith`, etc.)
- Enum sequence helpers used by class defaults

GHC.Base is `NoImplicitPrelude` and only imports `RHC.Prim`.  It
defines a local `mapInt`/`appendStr` instead of pulling
`Data.List.map`/`(++)` — keeps it self-contained while the list
combinators wait for their own boot module.

`lib/Prelude.hs` is now a thin shell: imports RHC.Prim,
Data.Function, GHC.Base, then defines only the list combinators
(`map`/`filter`/`(++)`/`head`/…), Maybe/Either helpers
(`maybe`/`fromMaybe`/`either`), `print`, `sum`/`product`/`replicate`/
`reverse`.  Re-exports everything via the explicit export list.

Driver / REPL wiring (mirrors the Data.Function step):

- `build.zig` bakes `ghc_base_path` (AOT) and embeds
  `ghc_base_source` (REPL); installs the file under
  `lib/ghc-internal/GHC/Base.hs`.
- `src/main.zig` boot_modules: RHC.Prim → Data.Function →
  GHC.Base → Prelude.
- `src/repl/session.zig` calls `loadBootModule` for it.

Plus a JIT correctness fix that surfaced once boot modules
multiplied: thread `Pipeline.grin_name_counter` through every
`translateProgram` call so cross-module GRIN helper names
(`_thunk_N`, lifted helpers) stay globally unique.  Without
this, RHC.Prim/Data.Function/GHC.Base/Prelude all emit `_thunk_17`
independently and ORC rejects the second module with
`Duplicate definition of symbol '_thunk_17'`.  Mirrors
`cmdBuild`'s `next_grin_name_counter`.

`zig build test --summary all` — 1216/1216 pass.
